fix: move Options to other component
