+++ src/components/atoms/Option/Option.types.ts
@@ -1,0 +1,12 @@
+type Option = {
+  value: string
+  label: string
+  img: string
+}
+
+export type OptionProps = {
+  option: Option
+  checked: boolean
+  onChange: (e: string) => void
+  name: string
+}
