+++ src/components/atoms/RadioInput/RadioInput.types.ts
@@ -1,0 +1,11 @@
+type RadioOption = {
+  label: string
+  value: string
+  img: string
+}
+
+export type RadioInputProps = {
+  label: string
+  name: string
+  options: RadioOption[]
+}
